apps/backend/requirements.txt
apscheduler==3.10.4
backtesting==0.3.3
ccxt==4.4.61
duckdb==1.1.3
fastapi==0.115.6
feedparser==6.0.11
httpx==0.28.1
mypy==1.14.1
polars==1.17.1
optuna==3.6.1
pydantic-settings==2.7.0
pyyaml==6.0.2
pytest==8.3.4
pytest-asyncio==0.25.0
python-dotenv==1.0.1
sqlmodel==0.0.22
uvicorn[standard]==0.34.0
vectorbt==0.26.2
